requirements.txt
tensorflow>=2.13.0
streamlit>=1.29.0
streamlit-drawable-canvas>=0.9.3
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.8.0
